+++ resources/themes/tabler/views/starter/templates/layouts/landing.blade.php
@@ -1,3 +1,4 @@
+@php($brandLogoUrl = $clientLogoUrl ?: asset('assets/tabler/static/logo-small.svg'))
 <!doctype html>
 <html lang="{{ str_replace('_', '-', app()->getLocale()) }}">
 
@@ -6,7 +7,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1, viewport-fit=cover">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>{{ $title ?? config('app.name') }} | {{ config('app.name') }}</title>
-    <link rel="shortcut icon" href="{{ asset('assets/tabler/static/logo-small.svg') }}">
+    <link rel="shortcut icon" href="{{ $brandLogoUrl }}">
     <link rel="stylesheet" href="{{ asset('assets/tabler/dist/css/tabler.min.css') }}?v={{ filemtime(public_path('assets/tabler/dist/css/tabler.min.css')) }}">
     <link rel="stylesheet" href="{{ asset('assets/tabler/css/tabler.css') }}?v={{ filemtime(public_path('assets/tabler/css/tabler.css')) }}">
     @includeIf('extensions.starter.layout.head')