+++ resources/themes/vuexy/views/starter/templates/components/company-brand.blade.php
@@ -1,0 +1,10 @@
+@include('starter.templates.layouts.brand', [
+    'url' => url('/'),
+    'navigate' => false,
+    'clientLogoUrl' => $clientLogoUrl,
+    'brandLogoUrl' => $clientLogoUrl ?: asset('assets/vuexy/img/branding/vuexy-mark.svg'),
+    'brandLogoAlt' => $clientLogoUrl ? ($clientName ?: config('app.name')) : config('app.name'),
+    'brandText' => config('app.name'),
+    'markClass' => 'vuexy-brand-mark-horizontal',
+    'showText' => ! $clientLogoUrl,
+])